ADMIN policies for movie CRUD

+++ app/controllers/movies_controller.rb
@@ -7,13 +7,17 @@ class MoviesController < ApplicationController
     @rating = Rating.create
   end
 
-  def show; end
+  def show;
+    authorize Movie
+  end
 
   def new
+    authorize Movie
     @movie = Movie.new
   end
 
   def create
+    authorize Movie
     @movie = Movie.new(movie_params)
     if @movie.save
       flash = 'The movie was successfully created'
@@ -24,9 +28,12 @@ class MoviesController < ApplicationController
     end
   end
 
-  def edit; end
+  def edit
+    authorize @movie
+  end
 
   def update
+    authorize @movie
     if @movie.update(movie_params)
       flash = 'Movie was successfuly updated'
       redirect_to root_path
@@ -37,6 +44,7 @@ class MoviesController < ApplicationController
   end
 
   def destroy
+    authorize @movie
     return flash = 'Movie was successfuly destroy' if @movie.destroy
 
     redirect_back(fallback_location: root_path)
